feat: Add initial CSS passthrough, swap to .njk templating
This commit adds an initial style.css file and the requisite passthrough
configuration for it to be detected and accessible to Eleventy. It also
removes the configuration to use .html files for templating in favour of
using .njk files directly.

The base HTML file now includes regular metadata and opengraph metadata
fields: Conor Meagher linked to his personal website's repository to
show examples of how precisely this works.